+++ src/editor/Editor.js
@@ -580,7 +580,40 @@ class Editor extends EditorStartup {
   /**
    * @returns {void}
    */
+  elementRenamed (win, renameObj) {
+    this.svgCanvas.runExtensions(
+      'elementRenamed',
+      /** @type {module:svgcanvas.SvgCanvas#event:ext_elementRenamed} */ {
+        renameObj
+      }
+    )
+  }
+
+  /**
+   * @returns {void}
+   */
+  afterClear (win) {
+    this.svgCanvas.runExtensions(
+      'afterClear'
+    )
+  }
+
+  /**
+   * @returns {void}
+   */
+  beforeClear (win) {
+    this.svgCanvas.runExtensions(
+      'beforeClear'
+    )
+  }
+
+  /**
+   * @returns {void}
+   */
   zoomDone () {
+    for (const el of this.svgCanvas.selectedElements) {
+      this.svgCanvas.selectorManager.requestSelector(el).resize()
+    }
     this.updateWireFrame()
   }
 
@@ -910,6 +943,18 @@ class Editor extends EditorStartup {
         this.hideSourceEditor()
       }
     }
+  }
+
+  /**
+   * @returns {void}
+   */
+  toggleDynamicOutput (e) {
+    this.configObj.curConfig.dynamicOutput = e.detail.dynamic
+    this.svgCanvas.setConfig(this.configObj.curConfig)
+    const $editorDialog = document.getElementById('se-svg-editor-dialog')
+    const origSource = this.svgCanvas.getSvgString()
+    $editorDialog.setAttribute('dialog', 'open')
+    $editorDialog.setAttribute('value', origSource)
   }
 
   /**