+++ src/editor/EditorStartup.js
@@ -192,6 +192,11 @@ class EditorStartup {
     )
     this.svgCanvas.bind('contextset', this.contextChanged.bind(this))
     this.svgCanvas.bind('extension_added', this.extAdded.bind(this))
+    this.svgCanvas.bind('elementRenamed', this.elementRenamed.bind(this))
+
+    this.svgCanvas.bind('beforeClear', this.beforeClear.bind(this))
+    this.svgCanvas.bind('afterClear', this.afterClear.bind(this))
+
     this.svgCanvas.textActions.setInputElem($id('text'))
 
     this.setBackground(this.configObj.pref('bkgd_color'), this.configObj.pref('bkgd_url'))
@@ -448,6 +453,8 @@ class EditorStartup {
     $id('se-svg-editor-dialog').addEventListener('change', function (e) {
       if (e?.detail?.copy === 'click') {
         this.cancelOverlays(e)
+      } else if (e?.detail?.dialog === 'dynamic') {
+        this.toggleDynamicOutput(e)
       } else if (e?.detail?.dialog === 'closed') {
         this.hideSourceEditor()
       } else {
@@ -543,6 +550,10 @@ class EditorStartup {
       if (this.configObj.curConfig.gridColor) {
         $editDialog.setAttribute('gridcolor', this.configObj.curConfig.gridColor)
       }
+
+      if (this.configObj.curConfig.dynamicOutput) {
+        $editDialog.setAttribute('dynamicoutput', true)
+      }
     }.bind(this))
 
     // zoom
@@ -592,7 +603,6 @@ class EditorStartup {
     })
     // run callbacks stored by this.ready
     await this.runCallbacks()
-    window.addEventListener('message', this.messageListener.bind(this))
   }
 
   /**
@@ -683,24 +693,6 @@ class EditorStartup {
     } catch (err) {
       // Todo: Report errors through the UI
       console.error(err)
-    }
-  }
-
-  /**
- * @param {PlainObject} info
- * @param {any} info.data
- * @param {string} info.origin
- * @fires module:svgcanvas.SvgCanvas#event:message
- * @returns {void}
- */
-  messageListener ({ data, origin }) {
-    const messageObj = { data, origin }
-    if (!this.extensionsAdded) {
-      this.messageQueue.push(messageObj)
-    } else {
-    // Extensions can handle messages at this stage with their own
-    //  canvas `message` listeners
-      this.svgCanvas.call('message', messageObj)
     }
   }
 }